refactor: consolidate jsx and transformjsx utils
